backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.24.3
scipy==1.11.4
torch==2.1.2
tensorflow==2.15.0
soundfile==0.12.1
librosa==0.10.1
boto3==1.34.14
resampy==0.4.2
python-multipart==0.0.6
websockets==12.0
localstack==3.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
aiohttp==3.9.1